fix(gemm): revert gfx1250 asm gemm interface to aiter/ops
