+++ aiter/__init__.py
@@ -97,6 +97,7 @@ else:
     from .ops.gemm_op_a8w8 import *
     from .ops.gemm_op_a16w16 import *
     from .ops.gemm_op_a4w4 import *
+    from .ops.gemm_op_a8w4 import *
     from .ops.batched_gemm_op_a8w8 import *
     from .ops.batched_gemm_op_bf16 import *
     from .ops.deepgemm import *
