+++ op_tests/test_f4gemm.py
@@ -23,6 +23,7 @@
 
 import argparse
 import itertools
+import sys
 
 import pandas as pd
 import torch
@@ -30,14 +31,7 @@ import torch
 import aiter
 from aiter import dtypes
 from aiter.jit.utils.chip_info import get_gfx_runtime as get_gfx
-from aiter.ops.asm.f4gemm import (
-    MXFP8_OUT_SCALE_BLOCK,
-    _gemm_a4w4o4_asm,
-    _gemm_a4w4o8_asm,
-    _gemm_mxfp4_asm,
-    _gemm_nvfp4_asm,
-    unpack_mxfp8_out_scale,
-)
+from aiter.ops.gemm_op_a4w4 import MXFP8_OUT_SCALE_BLOCK, unpack_mxfp8_out_scale
 from aiter.ops.shuffle import shuffle_scale_f4, shuffle_weight_f4
 from aiter.test_common import benchmark, checkAllclose, run_perftest
 from aiter.utility import fp4_utils
@@ -57,6 +51,35 @@ pd.set_option("display.width", 1000)
 SUPPORTED_GFX = ["gfx1250"]
 
 _OUT_DTYPE = {"bf16": dtypes.bf16, "fp8": dtypes.fp8, "fp4": dtypes.fp4x2}
+
+# gfx1250 F4GEMM .co is a persistent shader: it always launches PERSISTENT_TG
+# threadgroups regardless of problem size (must match the .co's WG_MAX).
+PERSISTENT_TG = 256
+
+
+def _report_active_tg(M, N, tile_m, tile_n, label):
+    """Warn when the persistent shader's TG slots aren't fully packed.
+
+    The .co always launches PERSISTENT_TG (256) threadgroups. The real work is
+    ceil(M/tile_m) * ceil(N/tile_n) tiles; when that isn't a multiple of 256 the
+    final wave leaves the leftover TG slots idle (wasted CUs) -> "poor perf".
+    (Moved here from the cpp dispatch so the report lives with the test.)
+    """
+    tg_m = (M + tile_m - 1) // tile_m
+    tg_n = (N + tile_n - 1) // tile_n
+    active_tg = tg_m * tg_n
+    wave_active = (
+        PERSISTENT_TG if active_tg % PERSISTENT_TG == 0 else active_tg % PERSISTENT_TG
+    )
+    info = (
+        f"{label}: active {wave_active}/{PERSISTENT_TG} TG "
+        f"({tg_m} M-tiles x {tg_n} N-tiles, tile_m={tile_m}, tile_n={tile_n})"
+    )
+    if active_tg % PERSISTENT_TG == 0:
+        aiter.logger.info("dispatch to %s", info)
+    else:
+        tag = "\033[31mpoor perf\033[0m" if sys.stderr.isatty() else "poor perf"
+        aiter.logger.warning("dispatch to %s - %s!", info, tag)
 
 PERF_SHAPES = [(16384, 16384, 16384)]
 FUNC_SHAPES = [
@@ -375,7 +398,7 @@ def test_gemm(
     if intype == "nvfp4":
 
         def run_asm(A, B, sA, sB, gA, gB):
-            return _gemm_nvfp4_asm(
+            return aiter.gemm_nvfp4_asm(
                 A,
                 B,
                 sA,
@@ -391,7 +414,7 @@ def test_gemm(
     else:
 
         def run_asm(A, B, sA, sB):
-            return _gemm_mxfp4_asm(
+            return aiter.gemm_mxfp4_asm(
                 A,
                 B,
                 sA,
@@ -426,6 +449,8 @@ def test_gemm(
     # the verbatim knl_name otherwise (kept in the table, see main()).
     actual_knl = knl_name if (knl_name and knl_name != "auto") else base
     ret = {"gfx": get_gfx(), "knl_name": actual_knl}
+    # F4GEMM tiles are always 256x256 (see f4gemm.csv). Report TG occupancy.
+    _report_active_tg(M, N, 256, 256, base)
     # Only a missing .co is reported as "not support"; any other failure (OOM,
     # memory fault, shape assert, ...) must propagate, not show as a green cell.
     # An explicit --knl-name that isn't in the cfg is a real error (typo / missing
@@ -461,9 +486,9 @@ def test_gemm(
             ret[f"{name} err"] = float("nan")
             ret[f"{name} result"] = "not support"
             continue
-        # Func-mode only: check the low-precision op contracts by outtype -- fp4 ->
-        # _gemm_a4w4o4_asm (single tensor [.,N//2]), fp8 -> _gemm_a4w4o8_asm
-        # ((data, scale) tuple). bf16 output is covered by run_asm. Not timed/tabled.
+        # Func-mode only: check the high-level op contracts by outtype -- bf16 ->
+        # gemm_a4w4 (single tensor), fp4 -> gemm_a4w4o4 (single tensor [.,N//2]),
+        # fp8 -> gemm_a4w4o8 ((data, scale) tuple). Not timed/tabled.
         if mode == "func":
             a4_kwargs = {"apreshuffle": bool(apre)}
             if intype == "nvfp4":
@@ -475,17 +500,23 @@ def test_gemm(
                 )
             args = (inp["A"], inp["B"], inp["sA"], inp["sB"])
             if out_fp8:
-                o, s = _gemm_a4w4o8_asm(*args, **a4_kwargs)
+                o, s = aiter.gemm_a4w4o8(*args, **a4_kwargs)
                 assert o.shape == out[0].shape and s.shape == out[1].shape, (
                     f"gemm_a4w4o8 shape mismatch: {tuple(o.shape)}/{tuple(s.shape)} "
                     f"vs {tuple(out[0].shape)}/{tuple(out[1].shape)}"
                 )
             elif out_fp4:
-                res = _gemm_a4w4o4_asm(*args, **a4_kwargs)
+                res = aiter.gemm_a4w4o4(*args, **a4_kwargs)
                 assert not isinstance(res, tuple), "gemm_a4w4o4 must return a tensor"
                 assert (
                     res.shape == out.shape
                 ), f"gemm_a4w4o4 shape mismatch: {tuple(res.shape)} vs {tuple(out.shape)}"
+            else:  # bf16
+                res = aiter.gemm_a4w4(*args, dtype=out_dtype, **a4_kwargs)
+                assert not isinstance(res, tuple), "gemm_a4w4 must return a tensor"
+                assert (
+                    res.shape == out.shape
+                ), f"gemm_a4w4 shape mismatch: {tuple(res.shape)} vs {tuple(out.shape)}"
         if out_fp4:
             # e2m1 is deterministic: compare dequantized values with zero
             # tolerance (exact fp4-code match). Borderline RNE ties may differ.
